moved r_a to wind.py, replaced u* with u(z) in r_a
